maybe_extract_raw

+++ datasets/common.py
@@ -26,10 +26,11 @@ num_cpus = multiprocessing.cpu_count()
 
 MANIFEST_FILE = "manifest.jsonl.gz"
 
+
 @dataclass
 class SpeechCorpus:
-    name:str
-    url:str
+    name: str
+    url: str
 
     @abstractmethod
     def build_audiofile2text(self, path) -> Dict[str, str]:
@@ -40,21 +41,18 @@ class SpeechCorpus:
     def get_corpora() -> List[SpeechCorpus]:
         raise NotImplementedError
 
-
-    def get_raw_zipfile(self,download_dir) -> str:
+    def get_raw_zipfile(self, download_dir) -> str:
         return maybe_download_compressed(self.name, download_dir, self.url)
 
-    def maybe_extract_raw(self,raw_zipfile,processed_dir):
+    def maybe_extract_raw(self, raw_zipfile, processed_dir):
         raw_extracted_dir = f"{processed_dir}/raw/{self.name}"
         maybe_extract(raw_zipfile, raw_extracted_dir, False)
         return raw_extracted_dir
 
 
-
-
 def process_write_manifest(processed_dir, file2utt, audio_conf: AudioConfig):
     os.makedirs(processed_dir, exist_ok=True)
-
+    failed = lambda x:x is None
     samples = tqdm(
         s._asdict()
         for s in process_with_threadpool(
@@ -63,7 +61,7 @@ def process_write_manifest(processed_dir, file2utt, audio_conf: AudioConfig):
                 process_build_sample, processed_folder=processed_dir, ac=audio_conf
             ),
             max_workers=2 * num_cpus,
-        )
+        ) if not failed(s)
     )
     data_io.write_jsonl(f"{processed_dir}/{MANIFEST_FILE}", samples)
 
@@ -76,23 +74,31 @@ class AudioConfig(NamedTuple):
 def process_build_sample(
     audio_file, text, processed_folder, ac: AudioConfig
 ) -> ASRSample:
+    try:
+        file_name, len_in_seconds, num_frames = process_audio(
+            audio_file, processed_folder, ac
+        )
+        asr_sample = ASRSample(file_name, text, len_in_seconds, num_frames)
+    except Exception:
+        print(f"failed to process {audio_file}")
+        asr_sample = None
+    return asr_sample
+
+
+def process_audio(audio_file, processed_folder, ac: AudioConfig):
     suffix = Path(audio_file).suffix
     assert audio_file.startswith("/")
     file_name = audio_file[1:].replace("/", "_").replace(suffix, f".{ac.format}")
     processed_audio_file = f"{processed_folder}/{file_name}"
-
     if ac.bitrate is not None:
         cmd = f"sox {audio_file} -C {ac.bitrate} {processed_audio_file}"
     else:
         cmd = f"sox {audio_file} {processed_audio_file}"
-
     exec_command(cmd)
-
     si, ei = torchaudio.info(processed_audio_file)
     num_frames = si.length / si.channels
     len_in_seconds = num_frames / si.rate
-
-    return ASRSample(file_name, text, len_in_seconds, num_frames)
+    return file_name, len_in_seconds, num_frames
 
 
 def maybe_download_compressed(local_filename, download_folder, url, verbose=False):
@@ -116,10 +122,11 @@ def maybe_download(localfile, url, verbose):
 
 
 def get_extract_process_zip_data(
-    audio_config:AudioConfig,
+    audio_config: AudioConfig,
     corpus: SpeechCorpus,
-    download_dir:str,
-    processed_dir:str,
+    download_dir: str,
+    processed_dir: str,
+    remove_raw_extract=True,
 ):
     raw_zipfile = corpus.get_raw_zipfile(download_dir)
     ac = f"{audio_config.format}{'' if audio_config.bitrate is None else '_' + str(audio_config.bitrate)}"
@@ -127,12 +134,13 @@ def get_extract_process_zip_data(
     processed_targz = f"{download_dir}/{corpus_folder_name}.tar.gz"
     if not os.path.isfile(processed_targz):
         processed_corpus_dir = os.path.join(processed_dir, corpus_folder_name)
-        raw_data_dir = corpus.maybe_extract_raw(raw_zipfile,processed_dir)
+        raw_data_dir = corpus.maybe_extract_raw(raw_zipfile, processed_dir)
         file2utt = corpus.build_audiofile2text(raw_data_dir)
         process_write_manifest(processed_corpus_dir, file2utt, audio_config)
         folder_to_targz(download_dir, processed_corpus_dir)
         print(f"wrote {processed_targz}")
-        shutil.rmtree(raw_data_dir)
+        if remove_raw_extract:
+            shutil.rmtree(raw_data_dir)
     else:
         print(f"found {processed_targz}")
         unzip(processed_targz, processed_dir)
